Branch load_fonts at the API, not at the file system
loadAppFonts and loadFont do nothing on the web: they print a warning and
return. The file reading behind them was split into readFileBytesSync,
listFilesWithExtension and readPackageNameFromPubspec, which put three
primitives in the public-ish surface to serve code that never runs there.

Branching the two functions themselves is the same conditional import one
level up. The io library keeps the file system work and loses the kIsWeb
checks, because it is only selected where dart:io exists. load_fonts.dart
holds the documentation once and forwards.

flutter_sdk is no longer reachable from spot.dart either, since only the io
library looks the SDK up, so its own io/web split goes away and it is a
plain dart:io library again.
