+++ analysis_options.yaml
@@ -8,6 +8,13 @@ analyzer:
     - "test/checks/**.dart"
     - "**/*.g.dart"
     - "counter_app/**.dart"
+    - build/**
+    - android/**
+    - ios/**
+    - web/**
+    - windows/**
+    - macos/**
+    - linux/**
 
 # You can customize the lint rules set to your own liking. A list of all rules
 # can be found at https://dart-lang.github.io/linter/lints/options/options.html
