+++ lib/src/screenshot/load_fonts_io.dart
@@ -1,0 +1,262 @@
+/// The dart:io implementation of the font loading API.
+///
+/// `package:spot/src/screenshot/load_fonts.dart` documents these functions and
+/// picks between this library and the web one.
+library;
+
+import 'dart:convert';
+import 'dart:io';
+
+import 'package:dartx/dartx_io.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:spot/src/flutter/flutter_sdk.dart';
+import 'package:spot/src/flutter/frame_clock.dart';
+
+Future<void>? _loadAppFontsFuture;
+
+/// {@macro spot.loadAppFonts}
+Future<void> loadAppFonts() async {
+  TestWidgetsFlutterBinding.ensureInitialized();
+  // Loading fonts is spot in use, usually from flutter_test_config.dart
+  // before any test ran, which is what makes the first test's frames counted
+  // from the very first one.
+  FrameClock.startCounting();
+  final existingFuture = _loadAppFontsFuture;
+  if (existingFuture != null) {
+    return existingFuture;
+  }
+
+  final future = _loadAppFontsOnce();
+  _loadAppFontsFuture = future;
+
+  try {
+    await future;
+  } catch (e, stackTrace) {
+    _loadAppFontsFuture = null;
+    Error.throwWithStackTrace(e, stackTrace);
+  }
+}
+
+Future<void> _loadAppFontsOnce() async {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
+  await TestAsyncUtils.guard<void>(() async {
+    // First we load the Roboto font from the Flutter SDK, which most Android apps use.
+    // In case the app defines a custom Roboto fontFamily it will be overwritten when
+    // loading the fonts from the manifest
+    await _loadMaterialFontsFromSdk();
+
+    // Load all fonts defined in the FontManifest.json file
+    await _loadFontsFromFontManifest();
+  });
+}
+
+/// {@macro spot.loadFont}
+Future<void> loadFont(String family, List<String> fontPaths) async {
+  if (fontPaths.isEmpty) {
+    return;
+  }
+
+  await TestAsyncUtils.guard<void>(() async {
+    final fontLoader = FontLoader(family);
+    for (final path in fontPaths) {
+      final decodedPath = Uri.decodeComponent(path);
+      try {
+        final file = File(decodedPath);
+        if (file.existsSync()) {
+          final Uint8List bytes = file.readAsBytesSync();
+          fontLoader.addFont(Future.value(bytes.buffer.asByteData()));
+        } else {
+          final data = rootBundle.load(decodedPath);
+          fontLoader.addFont(Future.value(data));
+        }
+      } catch (e, stack) {
+        debugPrint("Could not load font $decodedPath\n$e\n$stack");
+      }
+    }
+    // the fontLoader is unusable after calling load().
+    // No need to cache or return it.
+    await fontLoader.load();
+  });
+}
+
+/// Loads the Roboto/RobotoCondensed/MaterialIcons fonts from the executing Flutter SDK
+Future<void> _loadMaterialFontsFromSdk() async {
+  final root = flutterSdkRoot().absolute.path;
+
+  final materialFontsDir =
+      Directory('$root/bin/cache/artifacts/material_fonts/');
+
+  final fontFormats = ['.ttf', '.otf', '.ttc'];
+  final existingFonts = materialFontsDir
+      .listSync()
+      // dartfmt come on,...
+      .whereType<File>()
+      .where(
+        (font) => fontFormats.any((element) => font.path.endsWith(element)),
+      )
+      .toList();
+
+  final robotoFonts = existingFonts
+      .where((font) {
+        final name = font.name.toLowerCase();
+        return name.startsWith('Roboto-'.toLowerCase());
+      })
+      .map((file) => file.path)
+      .toList();
+  if (robotoFonts.isEmpty) {
+    debugPrint("Warning: No Roboto font found in SDK");
+  }
+  await loadFont('Roboto', robotoFonts);
+
+  final robotoCondensedFonts = existingFonts
+      .where((font) {
+        final name = font.name.toLowerCase();
+        return name.startsWith('RobotoCondensed-'.toLowerCase());
+      })
+      .map((file) => file.path)
+      .toList();
+  await loadFont('RobotoCondensed', robotoCondensedFonts);
+
+  final materialIcons = existingFonts
+      .where((font) {
+        final name = font.name.toLowerCase();
+        return name.startsWith('MaterialIcons-'.toLowerCase());
+      })
+      .map((file) => file.path)
+      .toList();
+  await loadFont('MaterialIcons', materialIcons);
+}
+
+/// Loads the fonts from the FontManifest.json file.
+///
+/// Fonts defined in an app are accessible via it family name "MyFont"
+/// Fonts defined in a package are accessible via "packages/myPackage/MyFont"
+///
+/// Because each app can also be a package, each font is available with both
+/// notations.
+/// This allows packages to access their own fonts also via
+/// "packages/myPackage/MyFont" like users of the package would.
+Future<void> _loadFontsFromFontManifest() async {
+  // The FontManifest.json file is generated by the Flutter build process
+  // located in /build/flutter_assets/FontManifest.json and bundled within the app
+  final binding = TestWidgetsFlutterBinding.instance;
+  final fontManifestContent =
+      await binding.runAsync(() => rootBundle.loadString('FontManifest.json'));
+  final json = jsonDecode(fontManifestContent!);
+  final fontManifest = _FontManifest.fromJson(json);
+
+  // The name of the package running the tests. Used to also expose the test
+  // target's own fonts under "packages/<self>/MyFont", matching how Flutter
+  // resolves a `package:` set on a TextStyle/IconData.
+  final thisPackageName = _readPackageNameFromPubspec();
+
+  for (final item in fontManifest.fontFamilies) {
+    final packageAsset =
+        item.assets.firstOrNullWhere((it) => it.startsWith('packages/'));
+    final packageName = packageAsset?.split('/')[1];
+
+    if (packageName == null) {
+      // font asset in pubspec.yaml references a file relative to the pubspec.yaml
+      // e.g. asset: lib/fonts/MyFont.ttf
+
+      // Make it accessible as "MyFont" for the app/package itself
+      await loadFont(item.family, item.assets);
+
+      // A package may reference its own fonts with `package: '<self>'`, which
+      // Flutter resolves to "packages/<self>/MyFont" at lookup time. Register
+      // that name too so the font is reachable both ways, mirroring the
+      // packages/ asset case below.
+      if (thisPackageName != null) {
+        await loadFont(
+          'packages/$thisPackageName/${item.family}',
+          item.assets,
+        );
+      }
+    } else {
+      // font uses the package notation, which resolves relative to the packages lib/* directory
+      // asset: packages/<packageName>/<somewhereInsideLib>/MyFont.ttf
+
+      // Make it accessible as "MyFont" to be used by the package itself
+      final fontFamilyName = item.family.split('/').last;
+      await loadFont(fontFamilyName, item.assets);
+      // and "packages/<packageName>/MyFont" so that other packages would reference it.
+      await loadFont('packages/$packageName/$fontFamilyName', item.assets);
+    }
+  }
+}
+
+/// Reads the `name:` field from the test target's pubspec.yaml.
+///
+/// `flutter test` creates a generated main.dart next to the package root, so
+/// pubspec.yaml is expected next to [Platform.script]. Returns null when no
+/// pubspec.yaml is found or it does not declare a name.
+String? _readPackageNameFromPubspec() {
+  final pubspec = File.fromUri(Platform.script.resolve('pubspec.yaml'));
+  if (!pubspec.existsSync()) {
+    return null;
+  }
+  final content = pubspec.readAsStringSync();
+  final match = RegExp(
+    '''^\\s*name\\s*:\\s*['"]?([a-zA-Z0-9_]+)['"]?''',
+    multiLine: true,
+  ).firstMatch(content);
+  return match?.group(1);
+}
+
+/// Parsed representation of the FontManifest.json file
+class _FontManifest {
+  final List<_FontManifestFontFamily> fontFamilies;
+
+  /// Represents a Flutter FontManifest
+  _FontManifest(this.fontFamilies);
+
+  /// Parses the FontManifest.json file
+  ///
+  /// Example:
+  /// ```json
+  ///  [
+  ///    {
+  ///      "family": "packages/app_font/Montserrat",
+  ///      "fonts": [
+  ///        {
+  ///          "asset": "packages/app_font/fonts/Montserrat-Regular.ttf"
+  ///        }
+  ///      ]
+  ///    }
+  ///  ]
+  /// ```
+  factory _FontManifest.fromJson(dynamic json) {
+    if (json is! List) {
+      throw const FormatException('FontManifest must begin with a List');
+    }
+    final List<_FontManifestFontFamily> fontFamilies = [];
+    for (final family in json) {
+      if (family is! Map) continue;
+      final familyName = family['family'];
+      if (familyName is! String) continue;
+      final List<String> assets = [];
+      final fonts = family['fonts'];
+      if (fonts is! List) continue;
+      for (final font in fonts) {
+        if (font is! Map) continue;
+        final asset = font['asset'];
+        if (asset is! String) continue;
+        // there are other values like weight and style, but those are ignored by Flutter
+        // https://github.com/flutter/website/issues/3591#issuecomment-521806077
+        assets.add(asset);
+      }
+      fontFamilies.add(_FontManifestFontFamily(familyName, assets));
+    }
+    return _FontManifest(fontFamilies);
+  }
+}
+
+class _FontManifestFontFamily {
+  final String family;
+  final List<String> assets;
+
+  _FontManifestFontFamily(this.family, this.assets);
+}
